improve performance of multimodule build by reusing compiler objects
