+++ src/main/java/sbt_inc/SbtIncrementalCompilers.java
@@ -6,35 +6,29 @@ package sbt_inc;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.*;
 import java.util.stream.Collectors;
 import org.apache.commons.exec.LogOutputStream;
 import org.apache.maven.plugin.logging.Log;
 import sbt.internal.inc.*;
-import sbt.internal.inc.FileAnalysisStore;
 import sbt.internal.inc.ScalaInstance;
 import sbt.util.Logger;
-import scala.Option;
-import scala.jdk.FunctionWrappers;
 import scala_maven.MavenArtifactResolver;
 import scala_maven.VersionNumber;
 import scala_maven_executions.Fork;
 import scala_maven_executions.ForkLogger;
-import xsbti.PathBasedFile;
-import xsbti.T2;
-import xsbti.VirtualFile;
 import xsbti.compile.*;
 
 public final class SbtIncrementalCompilers {
+
+  private static File COMPILER_BRIDGE_JAR = null;
+  private static SbtIncrementalCompiler SBT_INCREMENTAL_COMPILER = null;
+
   public static SbtIncrementalCompiler make(
       File javaHome,
       MavenArtifactResolver resolver,
       File secondaryCacheDir,
       Log mavenLogger,
-      File cacheFile,
-      CompileOrder compileOrder,
       VersionNumber scalaVersion,
       Collection<File> compilerAndDependencies,
       Collection<File> libraryAndDependencies,
@@ -43,29 +37,37 @@ public final class SbtIncrementalCompilers {
       List<File> forkBootClasspath)
       throws Exception {
 
-    ScalaInstance scalaInstance =
-        ScalaInstances.makeScalaInstance(
-            scalaVersion.toString(), compilerAndDependencies, libraryAndDependencies);
+    if (COMPILER_BRIDGE_JAR == null) {
+      if (mavenLogger.isInfoEnabled()) {
+        mavenLogger.info("Building compiler bridge JAR");
+      }
 
-    File compilerBridgeJar =
-        CompilerBridgeFactory.getCompiledBridgeJar(
-            scalaVersion, scalaInstance, secondaryCacheDir, resolver, mavenLogger);
+      ScalaInstance scalaInstance =
+          ScalaInstances.makeScalaInstance(
+              scalaVersion.toString(), compilerAndDependencies, libraryAndDependencies);
+
+      COMPILER_BRIDGE_JAR =
+          CompilerBridgeFactory.getCompiledBridgeJar(
+              scalaVersion, scalaInstance, secondaryCacheDir, resolver, mavenLogger);
+    } else {
+      if (mavenLogger.isInfoEnabled()) {
+        mavenLogger.info("Reusing compiler bridge JAR: " + COMPILER_BRIDGE_JAR.getAbsolutePath());
+      }
+    }
 
     if (jvmArgs == null || jvmArgs.length == 0) {
       return makeInProcess(
           javaHome,
-          cacheFile,
-          compileOrder,
-          scalaInstance,
-          compilerBridgeJar,
-          new MavenLoggerSbtAdapter(mavenLogger));
+          COMPILER_BRIDGE_JAR,
+          new MavenLoggerSbtAdapter(mavenLogger),
+          scalaVersion.toString(),
+          compilerAndDependencies,
+          libraryAndDependencies);
     } else {
       return makeForkedProcess(
           javaHome,
-          cacheFile,
-          compileOrder,
-          compilerBridgeJar,
-          scalaVersion,
+          COMPILER_BRIDGE_JAR,
+          scalaVersion.toString(),
           compilerAndDependencies,
           libraryAndDependencies,
           mavenLogger,
@@ -77,27 +79,33 @@ public final class SbtIncrementalCompilers {
 
   static SbtIncrementalCompiler makeInProcess(
       File javaHome,
-      File cacheFile,
-      CompileOrder compileOrder,
-      ScalaInstance scalaInstance,
       File compilerBridgeJar,
-      Logger sbtLogger) {
+      Logger sbtLogger,
+      String scalaVersion,
+      Collection<File> compilerAndDependencies,
+      Collection<File> libraryAndDependencies) {
 
-    Compilers compilers = makeCompilers(scalaInstance, javaHome, compilerBridgeJar);
-    AnalysisStore analysisStore = AnalysisStore.getCachedStore(FileAnalysisStore.binary(cacheFile));
-    Setup setup = makeSetup(cacheFile, sbtLogger);
-    IncrementalCompiler compiler = ZincUtil.defaultIncrementalCompiler();
-
-    return new InProcessSbtIncrementalCompiler(
-        compilers, analysisStore, setup, compiler, compileOrder, sbtLogger);
+    if (SBT_INCREMENTAL_COMPILER == null) {
+      try {
+        SBT_INCREMENTAL_COMPILER =
+            new InProcessSbtIncrementalCompiler(
+                javaHome,
+                compilerBridgeJar,
+                sbtLogger,
+                scalaVersion,
+                compilerAndDependencies,
+                libraryAndDependencies);
+      } catch (Exception e) {
+        throw new RuntimeException(e);
+      }
+    }
+    return SBT_INCREMENTAL_COMPILER;
   }
 
   private static SbtIncrementalCompiler makeForkedProcess(
       File javaHome,
-      File cacheFile,
-      CompileOrder compileOrder,
       File compilerBridgeJar,
-      VersionNumber scalaVersion,
+      String scalaVersion,
       Collection<File> compilerAndDependencies,
       Collection<File> libraryAndDependencies,
       Log mavenLogger,
@@ -108,7 +116,13 @@ public final class SbtIncrementalCompilers {
     List<String> forkClasspath =
         pluginArtifacts.stream().map(File::getPath).collect(Collectors.toList());
 
-    return (classpathElements, sources, classesDirectory, scalacOptions, javacOptions) -> {
+    return (classpathElements,
+        sources,
+        classesDirectory,
+        scalacOptions,
+        javacOptions,
+        compileOrder,
+        cacheFile) -> {
       try {
         String[] args =
             new ForkedSbtIncrementalCompilerMain.Args(
@@ -116,7 +130,7 @@ public final class SbtIncrementalCompilers {
                     cacheFile,
                     compileOrder,
                     compilerBridgeJar,
-                    scalaVersion.toString(),
+                    scalaVersion,
                     compilerAndDependencies,
                     libraryAndDependencies,
                     classpathElements,
@@ -179,68 +193,5 @@ public final class SbtIncrementalCompilers {
         throw new RuntimeException(e);
       }
     };
-  }
-
-  private static Compilers makeCompilers(
-      ScalaInstance scalaInstance, File javaHome, File compilerBridgeJar) {
-    ScalaCompiler scalaCompiler =
-        new AnalyzingCompiler(
-            scalaInstance, // scalaInstance
-            ZincCompilerUtil.constantBridgeProvider(scalaInstance, compilerBridgeJar), // provider
-            ClasspathOptionsUtil.auto(), // classpathOptions
-            new FunctionWrappers.FromJavaConsumer<>(noop -> {}), // onArgsHandler
-            Option.apply(null) // classLoaderCache
-            );
-
-    return ZincUtil.compilers(
-        scalaInstance, ClasspathOptionsUtil.boot(), Option.apply(javaHome.toPath()), scalaCompiler);
-  }
-
-  private static Setup makeSetup(File cacheFile, xsbti.Logger sbtLogger) {
-    PerClasspathEntryLookup lookup =
-        new PerClasspathEntryLookup() {
-          @Override
-          public Optional<CompileAnalysis> analysis(VirtualFile classpathEntry) {
-            Path path = ((PathBasedFile) classpathEntry).toPath();
-
-            String analysisStoreFileName = null;
-            if (Files.isDirectory(path)) {
-              if (path.getFileName().toString().equals("classes")) {
-                analysisStoreFileName = "compile";
-
-              } else if (path.getFileName().toString().equals("test-classes")) {
-                analysisStoreFileName = "test-compile";
-              }
-            }
-
-            if (analysisStoreFileName != null) {
-              File analysisStoreFile =
-                  path.getParent().resolve("analysis").resolve(analysisStoreFileName).toFile();
-              if (analysisStoreFile.exists()) {
-                return AnalysisStore.getCachedStore(FileAnalysisStore.binary(analysisStoreFile))
-                    .get()
-                    .map(AnalysisContents::getAnalysis);
-              }
-            }
-            return Optional.empty();
-          }
-
-          @Override
-          public DefinesClass definesClass(VirtualFile classpathEntry) {
-            return classpathEntry.name().equals("rt.jar")
-                ? className -> false
-                : Locate.definesClass(classpathEntry);
-          }
-        };
-
-    return Setup.of(
-        lookup, // lookup
-        false, // skip
-        cacheFile, // cacheFile
-        CompilerCache.fresh(), // cache
-        IncOptions.of(), // incOptions
-        new LoggedReporter(100, sbtLogger, pos -> pos), // reporter
-        Optional.empty(), // optionProgress
-        new T2[] {});
   }
 }
